requirements.txt
django==2.0.1
djangorestframework==3.7.7
celery==4.1.0
